# docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  model.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      finished_at:
        type: string
      id:
        type: string
      name:
        type: string
      priority:
        type: string
      project:
        type: string
      responsible_person:
        type: string
      state:
        type: string
    type: object
...
  model.User:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      registration_date:
        type: string
      role:
        type: string
    type: object
  model.UserInputResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      registration_date:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get User
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user in the system
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserInputResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Create User
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user from the system by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Delete User
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user from the system by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get User by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's details by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserInputResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Update User
      tags:
      - users
  /users/{id}/tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks assigned to a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get Tasks for User
      tags:
      - users
      - tasks
  /users/search:
    get:
      consumes:
      - application/json
      description: Search for a user in the system by their name or email
      parameters:
      - description: Username
        in: query
        name: name
        type: string
      - description: User Email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Search User
      tags:
      - users
swagger: "2.0"